backend/requirements.txt
aiofile==3.8.8
aiohttp-cors==0.7.0
aiohttp==3.8.6
aiosignal==1.3.1
python-socketio==5.11.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
caio==0.9.13
charset-normalizer==3.3.2
ecdsa==0.18.0
frozenlist==1.4.0
idna==3.4
loggate==1.10.0
multidict==6.0.4
pyasn1==0.5.0
python-jose==3.3.0
redis==5.0.1
rsa==4.9
six==1.16.0
soupsieve==2.5
yarl==1.9.2
pyyaml==6.0.1
SQLAlchemy==2.0.25
asyncpg==0.29.0
psycopg==3.1.17
alembic==1.13.1
pyaml==23.12.0
